requirements.txt
# This file works for Python 3.8.13
accelerate==0.20.3
adapter_transformers==3.2.1
numpy==1.24.3
opacus==1.4.0
scikit_learn==1.3.0
scipy==1.10.1
torch==2.0.1
tqdm==4.65.0
matplotlib==3.7.1
pandas==2.0.3